requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
requests>=2.28.0
python-dotenv>=0.19.0
openai>=1.0.0
beautifulsoup4>=4.11.0
httpx>=0.24.0
openpyxl>=3.0.0
supabase>=2.0.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
defeatbeta-api>=0.1.0
